Add more complete tests of roman and jwst datamodels
